Fix Failing Test Case
In the previous commit, we included a test case for checking the
ellipsis in `format`. While the test passed on the local system, it
failed on the remote system.

The reason is most likely the availability of the S3 method
`tbl_format_setup.my_tibble`. We already needed a special
`withr::with_environment` construct locally to make the test pass.
However, this does not properly guarantee availability on the remote
system.

Thus, we simply removed the specialized test case for a custom class and
only show the basic case where we expect an error, but not the case
where we do not expect an error.