fix: `print()` and `format()` pass on `...` to `tbl_format_setup()` again, as documented (#726)
* Fix Check on Ellipsis in `format`

The use of `rlang::check_dots_empty` in `format_tbl` threw an error
when arguments were passed via `...`. This behaviour seems unintended,
as we deliberately pass `...` down to `tbl_format_setup`.

In this fix, we replace `rlang::check_dots_empty` with
`rlang::check_dots_used` to ensure that all arguments passed to `format`
(or `print`, for that matter) are utilized. This change prevents
dangling arguments while allowing authors to add new arguments to their
print functions.

A dedicated test case was added to `test-tbl-format.R`. We used
`withr::with_environment` to ensure that the temporary
`tbl_format_setup.my_tibble` method can be properly dispatched.

* Fix Failing Test Case

In the previous commit, we included a test case for checking the
ellipsis in `format`. While the test passed on the local system, it
failed on the remote system.

The reason is most likely the availability of the S3 method
`tbl_format_setup.my_tibble`. We already needed a special
`withr::with_environment` construct locally to make the test pass.
However, this does not properly guarantee availability on the remote
system.

Thus, we simply removed the specialized test case for a custom class and
only show the basic case where we expect an error, but not the case
where we do not expect an error.

* Formatting

* Add snapshot test

---------

Co-authored-by: Kirill Müller <kirill@cynkra.com>